Refactor ChecklyService tests, enhance check result aggregation, and improve check status summarization. Updated test to run with dynamic account ID, added resultType to CheckResultAggregate, and optimized summarizeMultipleChecksStatus function with a currentDate parameter.
