+++ src/checkly/checklyclient.spec.ts
@@ -82,10 +82,8 @@ describe("ChecklyService", () => {
     expect(result).toBeDefined();
   });
 
-  it.skip("can run a check", async () => {
-    const result = await client.runCheck(
-      "e7608d1a-c013-4194-9da0-dec05d2fbabc",
-    );
+  it("can run a check", async () => {
+    const result = await client.runCheck(process.env.CHECKLY_ACCOUNT_ID!);
 
     expect(result).toBeDefined();
   });
