+++ src/slackbot/checkly.ts
@@ -1,6 +1,5 @@
 import { generateObject, generateText } from "ai";
 import { checkly } from "../checkly/client";
-import * as dataForge from "data-forge";
 import {
   analyseCheckFailureHeatMap,
   summariseMultipleChecksGoal,
@@ -33,6 +32,7 @@ import { generateHeatmap } from "../heatmap/generateHeatmap";
 import { createAccountSummaryBlock } from "./blocks/accountSummaryBlock";
 import { aggregateCheckResults } from "./check-result-slices";
 import { CheckResultsTimeSlice } from "./check-result-slices";
+import { summarizeCheckResultsToLabeledCheckStatus } from "./check-results-labeled";
 
 async function checkResultSummary(checkId: string, checkResultId: string) {
   const start = Date.now();
@@ -261,164 +261,6 @@ function toTimeBucket(startedAt: Date, bucketSizeInMinutes: number) {
   const sliceDate = new Date(startedAtDate);
   sliceDate.setMinutes(sliceMinute, 0, 0); // Set to start of 30-min bucket
   return sliceDate;
-}
-
-const hourlyFormatter = new Intl.DateTimeFormat("en-US", {
-  hour12: false,
-  month: "2-digit",
-  day: "2-digit",
-  hour: "2-digit",
-  minute: "2-digit",
-  timeZone: Intl.DateTimeFormat().resolvedOptions().timeZone,
-});
-
-export async function summarizeCheckResultsToLabeledCheckStatus(
-  aggregatedCheckResults: CheckResultAggregate[],
-) {
-  let df = new dataForge.DataFrame(aggregatedCheckResults);
-
-  const meanPassRatePerCheckAndLocation = df
-    .groupBy((row) => `${row.checkId}|${row.runLocation}`)
-    .select((group) => {
-      const passing = group.deflate((row) => row.passingCount).sum();
-      const degraded = group.deflate((row) => row.degradedCount).sum();
-      const failing = group.deflate((row) => row.errorCount).sum();
-      const passRateStdDev = group
-        .deflate((row) => row.passingCount / row.count)
-        .std();
-      const meanPassRate = passing / (passing + degraded + failing);
-
-      return {
-        checkId: group.first().checkId,
-        runLocation: group.first().runLocation,
-        meanPassRate,
-        passRateStdDev,
-      };
-    })
-    .reduce((acc, row) => {
-      acc.set(`${row.checkId}|${row.runLocation}`, {
-        meanPassRate: row.meanPassRate,
-        passRateStdDev: row.passRateStdDev,
-      });
-      return acc;
-    }, new Map<string, { meanPassRate: number; passRateStdDev: number }>());
-
-  log.debug("MEAN PASS RATE\n" + meanPassRatePerCheckAndLocation.toString());
-
-  interface CheckTimeSlice {
-    checkId: string;
-    runLocation: string;
-    startedAtBin: Date;
-    passing: number;
-    degraded: number;
-    failing: number;
-    passRate: number;
-    passRateDiff: number;
-    passRateStdDev: number;
-    degradedRate: number;
-    failRate: number;
-    cumSumPassRate?: number;
-  }
-
-  const checkIdLocationTimeSliceWithPassRateStdDev = df
-    .groupBy((row) => `${row.checkId}|${row.runLocation}|${row.startedAtBin}`)
-    .select((group): CheckTimeSlice => {
-      const checkId = group.first().checkId;
-      const runLocation = group.first().runLocation;
-      const { meanPassRate, passRateStdDev } =
-        meanPassRatePerCheckAndLocation.get(`${checkId}|${runLocation}`)!;
-
-      const count = group.deflate((row) => row.count).sum();
-      const passing = group.deflate((row) => row.passingCount).sum();
-      const degraded = group.deflate((row) => row.degradedCount).sum();
-      const failing = group.deflate((row) => row.errorCount).sum();
-
-      const passRate = passing / count;
-      const degradedRate = degraded / count;
-      const failRate = failing / count;
-
-      return {
-        checkId: group.first().checkId,
-        runLocation: group.first().runLocation,
-        startedAtBin: group.first().startedAtBin,
-        passing,
-        degraded,
-        failing,
-        passRate: passRate,
-        passRateDiff: passRate - meanPassRate,
-        passRateStdDev,
-        degradedRate,
-        failRate: failRate,
-      };
-    })
-    .inflate()
-    .withSeries<CheckTimeSlice & { cumSumPassRate: number }>(
-      "cumSumPassRate",
-      (s) => s.getSeries("passRateDiff").cumsum(),
-    )
-    .withSeries<
-      CheckTimeSlice & { cumSumPassRate: number; isCheckPoint: boolean }
-    >("isCheckPoint", (s) =>
-      s.deflate((row) =>
-        Math.abs(row.cumSumPassRate) > row.passRateStdDev * 2 ? 1 : 0,
-      ),
-    )
-    .withSeries<
-      CheckTimeSlice & {
-        cumSumPassRate: number;
-        isCheckPoint: boolean;
-        checkPointGroup: number;
-      }
-    >("checkPointGroup", (s) =>
-      new dataForge.Series([0]).concat(
-        s
-          .getSeries("isCheckPoint")
-          .rollingWindow(2)
-          .select((w) => {
-            const firstIsCheckPoint = w.first();
-            const lastIsCheckPoint = w.last();
-            return firstIsCheckPoint === lastIsCheckPoint ? 0 : 1;
-          })
-          .cumsum(),
-      ),
-    );
-
-  log.debug(
-    "CHECK ID LOCATION TIME SLICE WITH PASS RATE STD DEV\n" +
-      checkIdLocationTimeSliceWithPassRateStdDev.toString(),
-  );
-
-  const checksWithChangePoints = checkIdLocationTimeSliceWithPassRateStdDev
-    .groupBy((row) => `${row.checkId}|${row.runLocation}`)
-    .select((group) => {
-      const changePoints = group
-        .filter((row) => row.isCheckPoint)
-        .groupBy((row) => row.checkPointGroup)
-        .select((group) =>
-          group.orderBy((row) => Math.abs(row.cumSumPassRate)).last(),
-        )
-        .toArray();
-
-      return {
-        checkId: group.first().checkId,
-        runLocation: group.first().runLocation,
-        changePoints: changePoints.map((cp) => ({
-          timestamp: cp.startedAtBin.getTime(),
-          formattedTimestamp: hourlyFormatter.format(cp.startedAtBin),
-          severity:
-            cp.failRate > 0
-              ? "FAILING"
-              : cp.degradedRate > 0
-                ? "DEGRADED"
-                : "PASSING",
-        })),
-      };
-    })
-    .inflate();
-
-  log.debug("CHECKS WITH CHANGE POINTS\n" + checksWithChangePoints.toString());
-
-  return checksWithChangePoints;
 }
 
 async function checkSummaryData(
